ai/agents/pyproject.toml
[project]
name = "mcpeth"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.121.3",
    "fastmcp>=2.13.1",
    "langchain>=1.0.8",
    "langchain-groq>=1.0.1",
    "langchain-mcp-adapters>=0.1.13",
    "passlib>=1.7.4",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.38.0",
]
